+++ test/cases/options-postprocess/webpack.config.js
@@ -1,0 +1,43 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '/',
+    filename: 'assets/js/[name].[contenthash:8].js',
+  },
+
+  entry: {
+    index: './src/index.pug',
+  },
+
+  plugins: [
+    new PugPlugin({
+      extractCss: {
+        filename: 'assets/css/[name].[contenthash:8].css',
+      },
+      postprocess: (content, info, compilation) => {
+        return content.replace('REPLACE ME', 'The new replaced content.');
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+    ],
+  },
+};