+++ test/cases/responsive-images-many-duplicates/webpack.config.js
@@ -1,0 +1,66 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+  //mode: 'development',
+
+  resolve: {
+    alias: {
+      Images: path.join(__dirname, 'src/assets/images/'),
+      Styles: path.join(__dirname, 'src/assets/styles/'),
+    },
+  },
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    filename: 'assets/js/[name].[contenthash:8].js',
+  },
+
+  entry: {
+    index: './src/views/index.pug',
+  },
+
+  plugins: [
+    new PugPlugin({
+      extractCss: {
+        filename: (PathData) => {
+          // test auto publicPath for responsive images used in different scss files
+          if (PathData.chunk.name === 'main2') {
+            return 'assets/css2/subdir/[name].[contenthash:8].css';
+          }
+
+          return 'assets/css/[name].[contenthash:8].css';
+        },
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+
+      {
+        test: /\.(gif|png|jpe?g|ico|svg|webp)$/i,
+        type: 'asset/resource',
+        use: {
+          loader: 'responsive-loader',
+          options: {
+            name: 'assets/img/[name]-[width]w.[ext]',
+          },
+        },
+      },
+    ],
+  },
+};